Split converts into own files
Split converts into own files

Add copyright

Remove imports

Rename factory
