+++ genai-perf/genai_perf/inputs/converters/base_converter.py
@@ -1,0 +1,115 @@
+# Copyright 2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+#  * Neither the name of NVIDIA CORPORATION nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+# PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+# OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+import random
+from typing import Dict, List, Tuple
+
+from genai_perf.exceptions import GenAIPerfException
+from genai_perf.inputs.input_constants import ModelSelectionStrategy
+from genai_perf.inputs.inputs_config import InputsConfig
+
+
+class BaseConverter:
+    def convert(
+        self,
+        generic_dataset: Dict,
+        config: InputsConfig,
+    ) -> Dict:
+        raise NotImplementedError
+
+    def _select_model_name(
+        self,
+        config: InputsConfig,
+        index: int,
+    ) -> str:
+        if config.model_selection_strategy == ModelSelectionStrategy.ROUND_ROBIN:
+            return config.model_name[index % len(config.model_name)]
+        elif config.model_selection_strategy == ModelSelectionStrategy.RANDOM:
+            return random.choice(config.model_name)
+        else:
+            raise GenAIPerfException(
+                f"Model selection strategy '{config.model_selection_strategy}' is unsupported"
+            )
+
+    def _determine_json_feature_roles(
+        self, generic_dataset: Dict
+    ) -> Tuple[List[str], List[str], List[str]]:
+        SYSTEM_ROLE_LIST = ["system_prompt"]
+        USER_ROLE_LIST = ["question", "article"]
+        TEXT_INPUT_LIST = ["text_input"]
+
+        system_role_headers: List[str] = []
+        user_role_headers: List[str] = []
+        text_input_headers: List[str] = []
+
+        if "features" in generic_dataset.keys():
+            for feature in generic_dataset["features"]:
+                if feature in SYSTEM_ROLE_LIST:
+                    system_role_headers.append(feature)
+                if feature in USER_ROLE_LIST:
+                    user_role_headers.append(feature)
+                if feature in TEXT_INPUT_LIST:
+                    user_role_headers.append(feature)
+
+        assert (
+            system_role_headers is not None
+            or user_role_headers is not None
+            or text_input_headers is not None
+        )
+
+        return system_role_headers, user_role_headers, text_input_headers
+
+    def _create_new_text_input(
+        self,
+        header: str,
+        system_role_headers: List[str],
+        user_role_headers: List[str],
+        text_input_headers: List[str],
+        content: str,
+    ) -> str:
+        new_text_input = ""
+
+        if (
+            header in system_role_headers
+            or header in user_role_headers
+            or header in text_input_headers
+        ):
+            new_text_input = content
+
+        return new_text_input
+
+    def _add_new_text_input_to_json(
+        self, pa_json: Dict, index: int, new_text_input: str
+    ) -> Dict:
+        if new_text_input:
+            if pa_json["data"][index]["text_input"][0]:
+                pa_json["data"][index]["text_input"][0] = (
+                    pa_json["data"][index]["text_input"][0] + f" {new_text_input}"
+                )
+            else:
+                pa_json["data"][index]["text_input"][0] = new_text_input
+
+        return pa_json
