+++ genai-perf/genai_perf/inputs/converters/tensorrtllm_engine_converter.py
@@ -1,0 +1,86 @@
+# Copyright 2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+#  * Neither the name of NVIDIA CORPORATION nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+# PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+# OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+import random
+from copy import deepcopy
+from typing import Dict
+
+from genai_perf.inputs.converters.base_converter import BaseConverter
+from genai_perf.inputs.input_constants import (
+    DEFAULT_OUTPUT_TOKENS_MEAN,
+    DEFAULT_TENSORRTLLM_MAX_TOKENS,
+    EMPTY_JSON_IN_TENSORRTLLM_PA_FORMAT,
+)
+from genai_perf.inputs.inputs_config import InputsConfig
+
+
+class TensorRTLLMEngineConverter(BaseConverter):
+    def convert(
+        self,
+        generic_dataset: Dict,
+        config: InputsConfig,
+    ) -> Dict:
+        pa_json = deepcopy(EMPTY_JSON_IN_TENSORRTLLM_PA_FORMAT)
+
+        for index, entry in enumerate(generic_dataset["rows"]):
+            token_ids = config.tokenizer.encode(entry["text_input"])
+            pa_json["data"].append(
+                {
+                    "input_ids": {
+                        "content": token_ids,
+                        "shape": [len(token_ids)],
+                    },
+                    "input_lengths": [len(token_ids)],
+                    "request_output_len": [DEFAULT_TENSORRTLLM_MAX_TOKENS],
+                }
+            )
+
+            pa_json = self._add_optional_tags_to_trtllm_engine_json(
+                pa_json, index, config
+            )
+        return pa_json
+
+    def _add_optional_tags_to_trtllm_engine_json(
+        self,
+        pa_json: Dict,
+        index: int,
+        config,
+    ) -> Dict:
+        row = pa_json["data"][index]
+        if config.add_stream:
+            row["streaming"] = [True]
+        if config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN:
+            num_tokens = int(
+                random.gauss(config.output_tokens_mean, config.output_tokens_stddev)
+            )
+            row["request_output_len"] = [num_tokens]
+            if config.output_tokens_deterministic:
+                row["min_length"] = [num_tokens]
+
+        for key, value in config.extra_inputs.items():
+            row[key] = [value]
+
+        return pa_json
