+++ genai-perf/tests/test_output_format_converter.py
@@ -17,7 +17,9 @@ import random
 import pytest
 from genai_perf.inputs.input_constants import ModelSelectionStrategy, OutputFormat
 from genai_perf.inputs.inputs_config import InputsConfig
-from genai_perf.inputs.output_format_converter import OutputFormatConverterFactory
+from genai_perf.inputs.output_format_converter_factory import (
+    OutputFormatConverterFactory,
+)
 
 
 class TestOutputFormatConverter:
